fix: handle non-zero return code from Azure CLI
When running "om vm-lifecycle create-vm" on Azure, a problem has
recently occurred where querying a blob status using "az storage blob
show" has started to exit with a non-zero code and error "There is
currently a pending copy operation". Because of the error, the command
terminates. It's likely that this is because of a change in the backend
Azure API. As we are trying to determine whether the copy operation has
finished, it's safe to ignore this error and continue polling.

[#187303104](https://www.pivotaltracker.com/story/show/187303104)